+++ ctview/clip_raster.py
@@ -1,4 +1,8 @@
+from typing import List, Tuple
+
 from osgeo import gdal
+
+gdal.UseExceptions()
 
 
 def crop_raster(input_raster: str, output_raster: str, bounds: str):
@@ -19,9 +23,15 @@ def crop_raster(input_raster: str, output_raster: str, bounds: str):
     gdal.Translate(srcDS=input_raster, destName=output_raster, projWin=window)
 
 
-def clip_raster(input_raster: str, output_raster: str, bounds: str):
-    minX, minY, maxX, maxY = bounds[0][0], bounds[1][0], bounds[0][1], bounds[1][1]
-    RasterFormat = "GTiff"
-    gdal.Warp(output_raster, input_raster, format=RasterFormat, outputBounds=[minX, minY, maxX, maxY])
+def clip_raster(input_raster: str, output_raster: str, bounds: Tuple | List, raster_driver: str):
+    """Clip raster to the expected shape
 
-    return output_raster
+    Args:
+        input_raster (str): path to the input raster
+        output_raster (str): path to the output raster
+        bounds (Tuple | List): boundaries to clip to, as ((minx, maxx), (miny, maxy))
+        raster_driver (str): One of GDAL raster drivers formats
+        (cf. https://gdal.org/drivers/raster/index.html#raster-drivers)
+    """
+    minX, minY, maxX, maxY = bounds[0][0], bounds[1][0], bounds[0][1], bounds[1][1]
+    gdal.Warp(output_raster, input_raster, format=raster_driver, outputBounds=[minX, minY, maxX, maxY])