+++ ctview/utils_pdal.py
@@ -1,13 +1,25 @@
+import numpy as np
 import pdal
 
 
-def write_raster_class(input_points, output_raster, res):
-    """Generate a raster"""
+def write_raster_class(input_points: np.array, output_raster: str, res: float, raster_driver: str):
+    """Generate a raster with the contained classes by interpolation of all points that fall
+    in a resolution * sqrt(2) radius from the pixel centers.
+
+
+    Args:
+        input_points (np.array): Points of the input las (as read with a pdal.readers.las)
+        output_raster (str): path to the output raster
+        res (float): pixel size of the output raster
+        raster_driver (str): One of GDAL raster drivers formats
+        (cf. https://gdal.org/drivers/raster/index.html#raster-drivers)
+    """
+    """ """
     pipeline = pdal.Writer.gdal(
         filename=output_raster,
         resolution=res,
         dimension="Classification",
-        gdaldriver="GTiff",
+        gdaldriver=raster_driver,
     ).pipeline(input_points)
     pipeline.execute()