+++ RELEASE_NOTES.md
@@ -1,6 +1,9 @@
 # dev
-- generate buffer point cloud only once in the beggining of the algo
-- simplify config file
+- update pdal to 2.6+ 
+- refactor: 
+  - add gdal driver to config
+  - generate buffer point cloud only once in the beggining of the algo
+  - simplify config file
 - add option to gdal DEMprocessing to process hillshade on the sides as well (removes black lines)
 - density:
   - use same density map in main_ctView as in main_metadata